+++ resources/assets/js/modules/formatData.js
@@ -1,0 +1,24 @@
+// レスポンスデータを整形
+export default function formatData(response, categoryList) {
+  let date = new Date(response.date);
+  date =
+    date.getFullYear() + "." + (date.getMonth() + 1) + "." + date.getDate();
+
+  const responseCategory = response.categories.map(value => {
+    const targetList = categoryList.filter(category => {
+      return category.id === value;
+    });
+    return {
+      name: targetList[0].name,
+      url: targetList[0].url
+    };
+  });
+
+  return {
+    title: response.title.rendered,
+    url: response.link,
+    image: response._embedded["wp:featuredmedia"][0].source_url,
+    createddate: date,
+    categoryList: responseCategory
+  };
+}
